Add ceylon-format-region function and keybinding
Part of #8.

Error handling isn’t great (I managed to move the error output to a
separate buffer, but it still replaces the region with stdout, which in
case of error is empty), but a fix for that will probably be a bit
involved.

There are also numerous improvements to be made: fall back to entire
buffer if there is no region (the other part of #8); strip the trailing
newline if it wasn’t there in the original code; adjust indentation
(probably just copy the indentation of the beginning of the region and
prepend it to all following lines).