Fix initial indentation of first line
The initial indentation of the first line can be partially within the
region and partially outside of it. For example, suppose the first line
is indented by four spaces, and the region begins with the second space
(region-beginning is between the first and second space). The
indentation that must be added to all lines after the first one is four
spaces, but for the first line, one space was outside the region and was
therefore not removed by ceylon.formatter, so only three spaces need to
be added to that line.

Part of #8.