Use ceylon-mode-map and include in ceylon-mode documentation
Looks like I misunderstood mode-specific-map, so let’s use the
mode-dedicated map instead (as recommended in (elisp) Major Mode
Conventions), which already existed but I hadn’t noticed before. (But
move it to the end of the file, after the functions it uses.)

Also include the keymap in the documentation string for ceylon-mode.
\{KEYMAP} is special syntax to render the keymap as a table, which does
the right thing when the user customizes their table.