Improve regexp-opt instructions
With 'words instead of other non-nil PAREN argument, regexp-opt wraps
the regular expression in \<\> by itself. Also document how to run the
code better (I used to do this in the scratch window, but @xkr47
introduced me to M-:, which with C-u even inserts the value at point so
you don’t even have to copy+paste it from the echo area).