Add ceylon-compile command and C-c C-c keybinding
After all the preparatory work of the last commits, this doesn’t do much
more than stick the pieces together (which is good, because this
function is a lot less testable than the rest of them). We use the same
buffer as the ‘compile’ function and also put it in compilation-mode,
since the error information detection seems to work reasonably well on
the Ceylon compiler’s output as well.

As a special goodie, if the module name and source directory can’t be
detected, they default to “default” and “.” respectively, which means
that you can use ceylon-mode to just edit a .ceylon file in /tmp (or
anywhere) and use C-c C-c to compile it without having to bother setting
up a source directory structure or anything. (I’m also going to add a
ceylon-run function, with the same benefit, which will make this even
better.)

Closes #7.