Restore point after indendation
Record the column prior to the indentation, and then restore point to it
afterwards, adjusted for the change in indentation. This matches the
default behavior of many other programming modes; for users who don’t
like it, a customizable variable is added to disable this behavior and
keep point at the end of indentation as it was before.